Create demo module
