+++ package.json
@@ -16,6 +16,7 @@
     "axios": "^1.7.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.0",
     "react-speech-recognition": "^3.10.0",
     "regenerator-runtime": "^0.14.1",
     "rtcpeerconnection": "^8.4.0"
