+++ src/Develop/helper/index.ts
@@ -1,0 +1,1 @@
+export * from "./D_ID";
