+++ src/api.json
@@ -1,5 +1,5 @@
 {
-  "key": "dGVzdGFwaTcwMUBnbWFpbC5jb20:gWgKJAFetzzFP4T82Ol2j",
+  "key": "Ym9saXZhcmJyeWFuQGdtYWlsLmNvbQ:5JZiQbVAkk2Y6WrvFLxW5",
   "url": "https://api.d-id.com",
   "service": "talks"
 }
