--- src/helper/index.ts
@@ -1,1 +1,0 @@
-export * from "./D_ID";
