+++ src/main.tsx
@@ -1,6 +1,8 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.tsx";
+import Demo from "./Demo/Demo.tsx";
+import Develop from "./Develop/Develop.tsx";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
