+++ src/style.css
@@ -170,7 +170,7 @@ textarea {
 video {
   /* display: block; */
   border-radius: 50%;
-  background-image: url("./assets/emma_idle.png");
+  background-image: url("./assets/ALissa.jpeg");
   background-position: top;
   /* position: absolute; */
   background-size: contain;
